refactor noop conduit
